backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.1

# Image handling
Pillow==10.1.0
django-cloudinary-storage==0.3.0
cloudinary==1.36.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# API utilities
django-filter==23.5
drf-yasg==1.21.7

# Email
django-anymail==10.2

# Phone number validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.26

# Payment (PhonePe)
requests==2.31.0
pycryptodome==3.19.0

# Utilities
pytz==2023.3
